+++ R/bibentries.R
@@ -24,5 +24,18 @@ bibentries = c(
       series    = 	 "Proceedings of Machine Learning Research",
       address   = 	 "Cadiz, Spain",
       publisher =    "PMLR",
-      url       = "http://proceedings.mlr.press/v51/jamieson16.html")
+      url       = "http://proceedings.mlr.press/v51/jamieson16.html"),
+
+    li_2020 = bibentry("InProceedings",
+      key       = "li_2020",
+      title     = "A System for Massively Parallel Hyperparameter Tuning",
+      author    = "Liam Li and Kevin Jamieson and Afshin Rostamizadeh and Ekaterina Gonina and Jonathan Ben-tzur and Moritz Hardt and Benjamin Recht and Ameet Talwalkar",
+      booktitle = "Proceedings of Machine Learning and Systems",
+      editor    = "I. Dhillon and D. Papailiopoulos and V. Sze",
+      pages     = "230--246",
+      volume    = "2",
+      year      = "2020",
+      url       = "https://proceedings.mlsys.org/paper_files/paper/2020/hash/a06f20b349c6cf09a6b171c71b88bbfc-Abstract.html")
 )
+
+