+++ NEWS.md
@@ -1,3 +1,5 @@
+# mlr3hyperband (development version)
+
 # mlr3hyperband 0.6.0
 
 * compatibility; Work with new bbotk 1.0.0 and mlr3tuning 1.0.0