+++ NEWS.md
@@ -4,7 +4,7 @@
 
 # mlr3hyperband 0.6.0
 
-* compatibility; Work with new bbotk 1.0.0 and mlr3tuning 1.0.0
+* compatibility: Work with new bbotk 1.0.0 and mlr3tuning 1.0.0
 
 # mlr3hyperband 0.5.0