+++ NEWS.md
@@ -1,4 +1,4 @@
-# mlr3hyperband (development version)
+# mlr3hyperband 1.0.0
 
 * feat: Add `OptimizerAsyncSuccessiveHalving` optimizer.